Cleanup code

+++ Summa.Forms.WebApi/Controllers/FormController.cs
@@ -3,10 +3,12 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Summa.Forms.Models;
 using Summa.Forms.WebApi.Data;
+using Summa.Forms.WebApi.Extensions;
 using Summa.Forms.WebApi.Services;
 
 namespace Summa.Forms.WebApi.Controllers
@@ -17,13 +19,20 @@ namespace Summa.Forms.WebApi.Controllers
     public class FormController : ControllerBase
     {
         private readonly ApplicationDbContext _context;
+        private readonly IHttpContextAccessor _httpContextAccessor;
         private readonly IFormService _formService;
         private readonly IQuestionService _questionService;
         private readonly IResponseService _responseService;
 
-        public FormController(ApplicationDbContext context, IFormService formService, IQuestionService questionService, IResponseService responseService)
+        public FormController(
+            ApplicationDbContext context,
+            IHttpContextAccessor httpContextAccessor,
+            IFormService formService,
+            IQuestionService questionService,
+            IResponseService responseService)
         {
             _context = context;
+            _httpContextAccessor = httpContextAccessor;
             _formService = formService;
             _questionService = questionService;
             _responseService = responseService;
@@ -37,21 +46,35 @@ namespace Summa.Forms.WebApi.Controllers
         }
 
         [HttpPut("{formId}")]
-        public async Task<IActionResult> PutForm(Guid formId, [FromBody] Form form)
+        public async Task<IActionResult> PutForm(Guid formId, [FromBody] Form updated)
         {
-            if (formId != form.Id)
+            var subject = _httpContextAccessor.HttpContext.User.GetSubject().AsGuid();
+            var form = await _formService.GetByIdAsync(formId);
+            if (form.AuthorId != subject)
             {
-                return BadRequest();
+                return Unauthorized();
             }
 
-            var updated = await _formService.UpdateAsync(form);
-            return new JsonResult(updated, JsonSerializationConstants.SerializerOptions);
+            if (form.Id != updated.Id)
+            {
+                return BadRequest("Id's not match");
+            }
+
+            var value = await _formService.UpdateAsync(form, updated);
+            return new JsonResult(value, JsonSerializationConstants.SerializerOptions);
         }
 
         [HttpPost("{formId}/question")]
         public async Task<IActionResult> PostQuestion(Guid formId, [FromBody] Question question)
         {
+            var subject = _httpContextAccessor.HttpContext.User.GetSubject().AsGuid();
             var form = await _formService.GetByIdAsync(formId);
+
+            if (form.AuthorId != subject)
+            {
+                return Unauthorized();
+            }
+
             var created = await _formService.AddQuestionAsync(form, question);
 
             return new JsonResult(created, JsonSerializationConstants.SerializerOptions);
@@ -69,9 +92,16 @@ namespace Summa.Forms.WebApi.Controllers
         [HttpDelete("{formId}/question/{questionId}")]
         public async Task<IActionResult> DeleteQuestion(Guid formId, Guid questionId)
         {
+            var subject = _httpContextAccessor.HttpContext.User.GetSubject().AsGuid();
             var form = await _formService.GetByIdAsync(formId);
+
+            if (form.AuthorId != subject)
+            {
+                return Unauthorized();
+            }
+
             var question = await _questionService.GetByIdAsync(form, questionId);
-            
+
             await _formService.RemoveQuestionAsync(question);
 
             return NoContent();
@@ -80,7 +110,14 @@ namespace Summa.Forms.WebApi.Controllers
         [HttpPost("{formId}/question/{questionId}/option")]
         public async Task<IActionResult> PostOption(Guid formId, Guid questionId, [FromBody] QuestionOption option)
         {
+            var subject = _httpContextAccessor.HttpContext.User.GetSubject().AsGuid();
             var form = await _formService.GetByIdAsync(formId);
+
+            if (form.AuthorId != subject)
+            {
+                return Unauthorized();
+            }
+
             var question = await _questionService.GetByIdAsync(form, questionId);
             var created = await _questionService.AddOption(question, option);
 
@@ -90,7 +127,14 @@ namespace Summa.Forms.WebApi.Controllers
         [HttpDelete("{formId}/question/{questionId}/option/{optionId}")]
         public async Task<IActionResult> DeleteOption(Guid formId, Guid questionId, Guid optionId)
         {
+            var subject = _httpContextAccessor.HttpContext.User.GetSubject().AsGuid();
             var form = await _formService.GetByIdAsync(formId);
+
+            if (form.AuthorId != subject)
+            {
+                return Unauthorized();
+            }
+
             var question = await _questionService.GetByIdAsync(form, questionId);
             if (question.Options.Count <= 1)
             {
@@ -115,18 +159,16 @@ namespace Summa.Forms.WebApi.Controllers
         [HttpPost("{formId}/response")]
         public async Task<IActionResult> PostResponse(Guid formId, [FromBody] IEnumerable<QuestionAnswer> answers)
         {
-            var form = await _context.Forms
-                .Where(x => x.Id == formId)
-                .Include(x => x.Questions)
-                .AsNoTracking()
-                .FirstOrDefaultAsync();
+            var subject = _httpContextAccessor.HttpContext.User.GetSubject().AsGuid();
+            var form = await _formService.GetByIdAsync(formId);
+
+            if (form.AuthorId != subject)
+            {
+                return Unauthorized();
+            }
 
             var list = answers.ToList();
-            var valid = !list.Any(
-                answer => answer.QuestionId == Guid.Empty
-                          || form.Questions.All(x => x.Id != answer.QuestionId)
-            );
-
+            var valid = !list.Any(answer => answer.QuestionId == Guid.Empty || form.Questions.All(x => x.Id != answer.QuestionId));
             if (!valid)
             {
                 return BadRequest("Question identifier is missing");
